# Conflicts:
#	app/src/main/java/com/example/alphamobilecolombia/mvp/activity/UploadFileActivity.java

+++ app/src/main/java/com/example/alphamobilecolombia/services/implement/FileStorageJob.java
@@ -17,13 +17,14 @@ public class FileStorageJob {
     INotification _iNotification;
     int countStop = 3;
     int countTrans = 1;
-    public FileStorageJob(IUploadFilesPresenter iUploadFilesPresenter, IFileStorageService iFileStorageService, INotification iNotification){
+
+    public FileStorageJob(IUploadFilesPresenter iUploadFilesPresenter, IFileStorageService iFileStorageService, INotification iNotification) {
         _iUploadFilesPresenter = iUploadFilesPresenter;
         _iFileStorageService = iFileStorageService;
         _iNotification = iNotification;
     }
 
-    public void SendFilesToStorage(){
+    public void SendFilesToStorage() {
         List<FileStorage> fileStorages = _iFileStorageService.GetListForAll();
         List<String> listSubjectCredit = new ArrayList<>();
 
@@ -34,23 +35,23 @@ public class FileStorageJob {
                 }
             }
 
-            if(listSubjectCredit != null){
+            if (listSubjectCredit != null) {
                 for (String creditSubject : listSubjectCredit) {
-                    if(!creditSubject.equals("0"))
+                    if (!creditSubject.equals("0"))
                         SendFile(creditSubject);
                 }
             }
         }
     }
 
-    private boolean SendFile(String creditSubject){
+    private boolean SendFile(String creditSubject) {
         try {
             int documentNumber = 0;
             List<File> listUpload = new ArrayList<>();
             List<FileStorage> fileStoragesForCreditSubject = _iFileStorageService.GetListForCreditSubject(Integer.parseInt(creditSubject));
             for (FileStorage itemFile : fileStoragesForCreditSubject) {
                 documentNumber = itemFile.getDocumentNumber();
-                File newFile = new File(itemFile.getIdTypeFile(),itemFile.getName(),itemFile.isRequired(),itemFile.getNameType(),itemFile.isUpload(),itemFile.getFilePath(),true);
+                File newFile = new File(itemFile.getIdTypeFile(), itemFile.getName(), itemFile.isRequired(), itemFile.getNameType(), itemFile.isUpload(), itemFile.getFilePath(), true);
                 listUpload = new ArrayList<>();
                 listUpload.add(newFile);
 
@@ -63,23 +64,27 @@ public class FileStorageJob {
                 }
             }
 
-            //Cosnumir api de validación
-            boolean isCompletedCredit = true;
-            if (isCompletedCredit) {
-                _iFileStorageService.DeleteForCreditSubject(Integer.parseInt(creditSubject));
-                LocalNotification localNotification = new LocalNotification();
-                localNotification.setTitle("Proceso finalizado.");
-                localNotification.setMessage(creditSubject + " La solicitud de crédito para el cliente con número de documento " +documentNumber+" . Ha finalizado.");
-                _iNotification.ShowNotification(localNotification);
-            } else {
-                countTrans = countTrans +1;
-                if(countTrans <= countStop)
+            boolean isPendingFiles = _iUploadFilesPresenter.PendingFiles(creditSubject);
+            if (!isPendingFiles) {
+                //Cosnumir api de validación
+                boolean isCompletedCredit = _iUploadFilesPresenter.CompleteCredit(creditSubject);
+                if (isCompletedCredit) {
+                    _iFileStorageService.DeleteForCreditSubject(Integer.parseInt(creditSubject));
+                    LocalNotification localNotification = new LocalNotification();
+                    localNotification.setTitle("Proceso finalizado.");
+                    localNotification.setMessage(/*creditSubject + */" La solicitud de crédito para el cliente con número de documento " + documentNumber + " . Ha finalizado.");
+                    _iNotification.ShowNotification(localNotification);
+                }
+            }
+            else{
+                countTrans = countTrans + 1;
+                if (countTrans <= countStop)
                     SendFile(creditSubject);
             }
 
+
             return true;
-        }
-        catch (Exception ex){
+        } catch (Exception ex) {
             ex.printStackTrace();
             return false;
         }
